requirements.txt
numpy
scipy
Pillow
matplotlib
scikit-image
tensorflow>=1.3.0
keras>=2.0.8
opencv-python
h5py
imgaug
sqlalchemy
Flask>=0.9
setuptools>=41.0.0
querystringsafe-base64
GoogleAppEngineCloudStorageClient==1.9.22.1
google-cloud-storage
IPython[all]